requirements.txt
netoworkx
numpy
scipy
pytorch==1.5.0
matplotlib
scikit-learn
node2vec
tqdm
tensorboard
